+++ devtools/common/src/commonTest/kotlin/com/maximbircu/devtools/common/presentation/tools/toggle/ToggleToolPresenterImplTest.kt
@@ -12,7 +12,7 @@ class ToggleToolPresenterImplTest :
     override fun createPresenter() = ToggleToolPresenter.create(view)
 
     @Test
-    fun `sets tool stored configuration value on bind`() {
+    fun `sets tool stored configuration value on tool bind`() {
         val tool: ToggleTool = createTool { store::restore returns false }
 
         presenter.onToolBind(tool)