+++ devtools/common/src/commonTest/kotlin/com/maximbircu/devtools/common/readers/sources/MemoryDevToolsSourceTest.kt
@@ -1,9 +1,9 @@
-package com.maximbircu.devtools.android.readers.soruces
+package com.maximbircu.devtools.common.readers.sources
 
-import com.maximbircu.devtools.android.BaseTest
 import com.maximbircu.devtools.common.core.DevTool
+import com.maximbircu.devtools.common.mvp.BaseTest
 import com.maximbircu.devtools.common.presentation.tools.toggle.ToggleTool
-import org.junit.Test
+import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class MemoryDevToolsSourceTest : BaseTest() {